Move browse code under repository
